fix wrong PDF report bug
For some reason (this fixes it, but I don't yet know why), exporting some reports would always result in the first servicejob for the customer. Maybe it has something to do with the '/' in the accountname, I don't know. This fixes it but it needs more work since requiring the crm ID in the filename is not ideal.